+++ components/ui/common/RequireMark.tsx
@@ -1,0 +1,8 @@
+/**
+ * 必須マークの部品コンポーネント
+ */
+export default function RequireMark() {
+    return (
+        <span className="text-xs font-black text-white bg-pink-500 px-1.5 py-0.5 rounded-md">必須</span>
+    );
+}
